requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
Pillow>=10
watchdog>=4
python-multipart>=0.0.9
